+++ src/ThumbnailNavigation.h
@@ -1,0 +1,8 @@
+/* Copyright 2026 the SumatraPDF project authors (see AUTHORS file).
+   License: GPLv3 */
+
+struct MainWindow;
+struct WindowTab;
+
+void ShowThumbnailNavigation(MainWindow* win);
+void FreeThumbnailNavigationCache(WindowTab* tab);